// src/app/api/model/dividend-search-request.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DividendSearchRequest { 
    tickers?: Array<string> | null;
    dividend_type?: Array<DividendSearchRequestDividendTypeEnum> | null;
    frequency?: number | null;
    ex_dividend_date_from?: string | null;
    ex_dividend_date_to?: string | null;
    record_date_from?: string | null;
    record_date_to?: string | null;
    declaration_date_from?: string | null;
    declaration_date_to?: string | null;
    pay_date_from?: string | null;
    pay_date_to?: string | null;
    cash_amount_min?: number | null;
    cash_amount_max?: number | null;
    page?: number;
    page_size?: number;
}
export enum DividendSearchRequestDividendTypeEnum {
    CD = 'CD',
    SC = 'SC',
    LT = 'LT',
    ST = 'ST'
};