// src/app/api/api/dividend-search.service.ts
export interface SearchDividendsApiDividendsSearchPostRequestParams {
    dividendSearchRequest: DividendSearchRequest;
}